backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
supabase==2.10.0
e2b-code-interpreter==1.0.3
openai==1.57.0
# For handling file uploads if needed
python-multipart==0.0.19
